--- style.css
@@ -1,162 +1,0 @@
-:root{
-    --line-color:rgba(255,255,255,0.1);
-    --border-color:rgba(255,255,255,0.1);
-    --text-color:rgba(255,255,255,0.8);
-}
-body{
-    width: 100%;
-    height: 100vh;
-}
-#background-img{
-    height: 100vh;
-    background: url(/images/main-background.jpg) no-repeat;
-    background-size: 100% 100%;
-    background-position: center;
-    background-blend-mode: multiply;
-}
-.menuLine-inner{
-    width: 15%;
-    height: 100vh;
-    float: left;
-}
-#main-back{
-    width: 50%;
-    margin: 55px auto 0px auto;
-    position: relative; left: 0; top: 0;
-    cursor: pointer;
-}
-#main-back::before{
-    display: block;
-    content: "";
-    position: absolute; left: 50%; top: -60px;
-    margin-left: -1px;
-    width: 2px;
-    height: 40px;
-    background-color: var(--line-color);
-}
-#main-back::after{
-    display: block;
-    content: "";
-    position: absolute; left: 50%; top: 40px;
-    margin-left: -1px;
-    width: 2px;
-    height: calc(100vh - 95px);
-    background-color: var(--line-color);
-}
-#main-back div{
-    width: 46px;
-    height: 3px;
-    background-color: #f1f1f1;
-    border-radius: 2px;
-    margin: 0 auto 8px auto;
-}
-.section-wrap{
-    width: 60%;
-    float: left;
-}
-header{
-    width: 100%;
-}
-.menu-list{
-    display: flex;
-}
-.menu-list li{
-    width: 13%;
-    padding: 50px 0px;
-    font-size: 17px;
-    font-weight: 400;
-}
-section{
-
-}
-.section-inner{
-    padding: 4% 35% 0px 0px;
-}
-#detail-title{
-    font-size: 70px;
-    margin-bottom: 17px;
-    font-weight: 700;
-}
-.detail-overview{
-    font-size: 15px;
-    font-weight: 300;
-    line-height: 25px;
-    margin-bottom: 7%;
-    word-break: break-all;
-}
-.learn-more{
-    font-size: 17px;
-    position: relative; left: 0; top: 0;
-}
-.learn-more::after{
-    display: block;
-    content: "";
-    width: 154px;
-    height: 18px;
-    background: url(/images/learn-arrow.png) no-repeat;
-    position: absolute; left: 150px; top: 50%;
-    margin-top: -9px;
-}
-dl{
-    padding-top: 24%;
-}
-dt{
-    font-size: 22px;
-    color: var(--text-color);
-    font-weight: 500;
-    padding-bottom: 15px;
-}
-dd{
-    font-size: 20px;
-    font-weight: 300;
-}
-aside{
-    width: 25%;
-    height: 100vh;
-    float: right;
-    box-sizing: border-box;
-    border-left: solid 1px var(--border-color);
-}
-.sns-wrap{
-    
-}
-.sns{
-    display: flex;
-    padding: 45px;
-    border-bottom: solid 1px var(--border-color);
-}
-.sns-list{
-    display: flex;
-    width: 100%;
-    justify-content: space-around;
-
-}
-.sns-list li{
-}
-.followUs{
-    font-size: 25px;
-    font-weight: 500;
-    letter-spacing: -.5px;
-    color: var(--text-color);
-}
-.fa{
-    font-size: 20px;
-    color: var(--text-color);
-}
-.movieInfo-list{
-    padding: 48px 0px 0px 57px
-}
-.movieInfo-list li{
-    margin-bottom: 40px;
-}
-.movieInfo-title{
-    font-size: 20px;
-    font-weight: 500;
-    letter-spacing: .5px;
-    margin-bottom: 15px;
-    color: var(--text-color);
-}
-.movieInfo-info{
-    font-size: 15px;
-    letter-spacing: -.5px;
-}
